requirements.txt
boto3
python-twitter
motionless
reverse_geocoder
